DataAugmentation

DataAugmentation.AbstractAffineTransform

DataAugmentation.AbstractResizedTransform

DataAugmentation.AbstractTransform

DataAugmentation.AffineTransform

DataAugmentation.ComposedAffineTransform

DataAugmentation.CroppedAffineTransform

DataAugmentation.FlipX

DataAugmentation.FlipY

DataAugmentation.Image

DataAugmentation.Item

DataAugmentation.ItemWrapper

DataAugmentation.Keypoints

DataAugmentation.Label

DataAugmentation.Pipeline

DataAugmentation.Rotate

DataAugmentation.Tensor

DataAugmentation.Scale

DataAugmentation.applytfm

DataAugmentation.applytfmkeypoint

DataAugmentation.itemdata

DataAugmentation.pickfromrange

DataAugmentation.AbstractAffineTransform — Type

AbstractAffineTransform

Abstract supertype for affine transformations

Supporting custom AbstractAffineTransforms and Items

AbstractAffineTransform interface

gettfmmatrix(t::MyAffineTransform, getbounds(item), getparam(t)) should return a transformation matrix (s. CoordinateTransformations.jl)

Item interface

getbounds(item::MyItem)::Tuple returns the spatial bounds of an item, e.g. size(img) for an image array

applytfm(item::MyItem, tfm)::MyItem applies transformation matrix tfm (constructed with gettfmmatrix) to item and returns an item of the same type

DataAugmentation.AbstractTransform — Type

AbstractTransform

Abstract type for all transforms

AbstractTransform interface

TODO: Document

DataAugmentation.AffineTransform — Type

AffineTransform

Applies static transformation matrix tfm to an item

DataAugmentation.CroppedAffineTransform — Type

CroppedAffineTransform(transform, croptransform)

Applies an affine transform and crops with croptransform, such that getbounds(t(item)) == getcrop(croptransform)

This wrapper leads to performance improvements when warping an image, since only the indices within the bounds of crop need to be evaluated.

Composing any AbstractAffineTransform with a CropTransform constructs a CroppedAffineTransform.

DataAugmentation.FlipX — Type

FlipX

DataAugmentation.FlipY — Type

FlipY

DataAugmentation.Image — Type

Image

Item for an Image

DataAugmentation.Item — Type

`Item`

Abstract data container

TODO: document Item interface

DataAugmentation.ItemWrapper — Type

`ItemWrapper`

Abstract type for wrapper items

TODO: document ItemWrapper interface

DataAugmentation.Keypoints — Type

Keypoints

DataAugmentation.Label — Type

Label

DataAugmentation.Pipeline — Type

Pipeline

A Pipeline combines several transformations and applies them sequentially

DataAugmentation.Rotate — Type

RotateTransform(angle)
RotateTransform(angles)

Rotates an Item by angle degrees counter-clockwise using an affine transformation. If a vector or range angles of degrees is given, picks a random element.

See also Rotate90, Rotate180, Rotate270

DataAugmentation.Tensor — Type

Tensor

DataAugmentation.Scale — Method

Scale

DataAugmentation.itemdata — Method

itemdata(item::Item)

TODO: document

DataAugmentation.AbstractResizedTransform — Type

DataAugmentation.ComposedAffineTransform — Type

ComposedAffineTransform(transforms)

Composes several affine transformations.

Due to associativity of affine transformations, the transforms can be combined before applying, leading to large performance improvements.

Chaining (|>) multiple AbstractAffineTransformations automatically creates a ComposedAffineTransform.

DataAugmentation.applytfm — Method

applytfm(item, tfm, crop = nothing)

Applies affine transformation matrix tfm to item, optionally cropping to window of size crop

DataAugmentation.applytfmkeypoint — Method

applytfmkeypoint(k, tfm)

Applies affine transformation matrix tfm to a single keypoint k

DataAugmentation.pickfromrange — Method

pickfromrange(rng::AbstractRange, rnd::AbstractFloat)

Given a number 0 <= rnd <= 1 that's generated by rand(), pick the element in range rng that is closest to it relative to rng's length