DataAugmentation.AbstractAffine

DataAugmentation.Affine

DataAugmentation.ComposedAffine

DataAugmentation.CroppedAffine

DataAugmentation.Either

DataAugmentation.Either

DataAugmentation.FlipX

DataAugmentation.FlipY

DataAugmentation.Identity

DataAugmentation.Lambda

DataAugmentation.Pipeline

DataAugmentation.Rotate

DataAugmentation.Rotate180

DataAugmentation.Rotate270

DataAugmentation.Rotate90

DataAugmentation.Scale

DataAugmentation.apply

DataAugmentation.applyaffine

DataAugmentation.applyaffine

DataAugmentation.applystep

DataAugmentation.compose

DataAugmentation.pickfromrange

DataAugmentation.AbstractAffine — Type

AbstractAffineTransform

Abstract supertype for affine transformations

Supporting custom AbstractAffineTransforms and Items

AbstractAffineTransform interface

getaffine(t::MyAffineTransform, getbounds(item), getparam(t)) should return a transformation matrix (s. CoordinateTransformations.jl)

Item interface

getbounds(item::MyItem)::Tuple returns the spatial bounds of an item, e.g. size(img) for an image array

applyaffine(item::MyItem, A)::MyItem applies transformation matrix A (constructed with getaffine) to item and returns an item of the same type

DataAugmentation.Affine — Type

Affine

Applies static transformation matrix tfm to an item

DataAugmentation.ComposedAffine — Type

ComposedAffine(transforms)

Composes several affine transformations.

Due to associativity of affine transformations, the transforms can be combined before applying, leading to large performance improvements.

composeing multiple AbstractAffineTransformations automatically creates a ComposedAffine.

DataAugmentation.CroppedAffine — Type

CroppedAffine(transform, croptransform)

Applies an affine transform and crops with croptransform, such that getbounds(t(item)) == getcrop(croptransform)

This wrapper leads to performance improvements when warping an image, since only the indices within the bounds of crop need to be evaluated.

composeing any AbstractAffine with a CropTransform constructs a CroppedAffine.

DataAugmentation.Either — Type

Either(transformations, probabilities)

Chooses one of transformations with probability in probabilities

DataAugmentation.Either — Method

Either(transform, p)

Applies tfm with probability p

DataAugmentation.FlipX — Type

FlipX()

Flips an item along the x-axis

DataAugmentation.FlipY — Type

FlipY()

Flips an Item along the y-axis

DataAugmentation.Identity — Type

Identity()

Does not transform an item

DataAugmentation.Lambda — Type

Lambda(f)

Applies function f to an Item

DataAugmentation.Pipeline — Type

Pipeline

A Pipeline combines several transformations and applies them sequentially

DataAugmentation.Rotate — Type

RotateTransform(angle)
RotateTransform(angles)

Rotates an Item by angle degrees counter-clockwise using an affine transformation. If a vector or range angles of degrees is given, picks one at random.

See also Rotate90, Rotate180, Rotate270 for more efficient non-affine versions for those specific angles.

DataAugmentation.Rotate180 — Type

Rotate180()

Rotate clock-wise by 180 degress

Faster than Rotate for array types.

DataAugmentation.Rotate270 — Type

Rotate270()

Rotate clock-wise by 270 degress

Faster than Rotate for array types.

DataAugmentation.Rotate90 — Type

Rotate90()

Rotate clock-wise by 90 degress

Faster than Rotate for array types.

DataAugmentation.Scale — Method

Scale

DataAugmentation.apply — Method

apply(tfm, item)
apply(tfm, items)

Apply transform tfm to a single item or a tuple of items. If given a tuple, the same random state is used to transform the individual items.

DataAugmentation.applyaffine — Function

applyaffine(item::Item, A, crop = nothing)

Applies an affine transformation A to item, optionally cropping to tuple crop.

DataAugmentation.applyaffine — Method

applyaffine(item, tfm, crop = nothing)

Applies affine transformation matrix tfm to item, optionally cropping to window of size crop

DataAugmentation.applystep — Method

applystep(tfm, items, names)

Applies tfm inplace to items in items given by keys names.

DataAugmentation.compose — Method

compose(tfm1, tfm2)
tfm1 |> tfm2

Combines two Transforms. The default behavior is to group them in a Pipeline that applies them sequentially.

Custom methods can be used to implement performance improvements, for example merging two affine transformations.

DataAugmentation.pickfromrange — Method

pickfromrange(rng::AbstractRange, rnd::AbstractFloat)

Given a number 0 <= rnd <= 1 that's generated by rand(), pick the element in range rng that is closest to it relative to rng's length